+++ src/renderer/tools/StartOnBoot.js
