+++ src/renderer/tools/api/Public.js
@@ -1,0 +1,14 @@
+import {Ajax,severAddress} from "./request";
+export default {
+    Check(url,callback,error){
+        Ajax({
+            url:url,
+            data:[],
+            success:callback,
+            error:error
+        })
+    },
+    VerifyCode(){
+        return severAddress()+"/service/verifyCode"+'?'+Math.random();
+    }
+}
