+++ src/renderer/tools/api/index.js
@@ -1,0 +1,7 @@
+import User from "./User";
+import Disk from "./Disk";
+import Public from "./Public";
+import LocalFile from "./LocalFile";
+export default {
+    User,Disk,Public,LocalFile
+}
